ooxml: preserve font theme color property
The property is called w:themeColor and it belongs to the w:color tag
inside run properties.

On import, the themeColor prop is stored in the character interop
grab bag together with the val prop.

On export, the current color of a text run is compared with the
stored val prop to know if the color has been changed during the
document edition. If it hasn't changed, the themeColor property
must be written to the document; if it has, it must not be written to
avoid overwriting the new color.

Also added a unit test for this property.

Change-Id: Icc95ee023aecc741adcba23d23206aadd6c30e1f