fdo#72452: ODF import/export: fix handling of style:page-number
Class XMLNumberWithAutoInsteadZeroPropHdl (which appears to be used only
for this attribute) needs to be adapted to the change that
"PageNumberOffset" value 0 is no longer invalid; use "void" value for
invalid instead, which appears more appropriate anyway.

Unfortunately the type of style:page-number is positiveInteger so
writing 0 would be invalid; write "auto" instead for now.

Change-Id: I9621ea201fd928087b863c562607c3d77a3b0269