Fix crash on Android
Fallout from f9ab1e9cccd40cd62a7e7b3c001973df8bdba216 ("Convert code that
calls OUString::getStr()[] to use the [] operator"). The code was a bit silly
and "hand-optimized" in the first place so no wonder a bug was introduced
accidentally in that commit.

Make it simpler and more obvious.

Change-Id: Idafb1c0741289e915dfbaa108383a926321764d2