Introduce DPI scale factor for Hi-DPI displays.
This is supposed to stay 1 (no scale) for printers and default virtual
devices, but should be set accordingly (2 or 3) for windows and virtual
devices derived from windows.

Various VCL widgets or paint operations should incrementally become aware of
this, and changed so that they draw nicely on the Hi-DPI displays.  This patch
only introduces the behavior for waved lines.

The default is currently being set depending on the DPI setting only; could be
changed to a more clever way if necessary.

Change-Id: I71118f9ab6b64028d1eeee76e860e999d5cd9d19