Revert "Avoid unnecessary library re-building in the DISABLE_DYNLOADING case"
Unfortunately doing it like that breaks the case where as part of
building a dependent library some headers are generated and those
headers are then needed when compiling a depending library. Case in
point: sqlbison.hxx.

(Sure, for that particular case, as such it is fairly likely that we
don't want any of the database connectivity code at all in any iOS or
Android case, but working on the --disable-database-connectivity
configure switch is another task...)

This reverts commit 7a63503c4819a97bbc3974eac136d239a94e6c69.

Change-Id: I8dd1f260160c90473ecf12ce862f37df9767fa19