ooxml: Preserve paragraph theme fill attribute
The theme fill attribute and the original fill color are saved to the
paragraph grab bag during the import.

On export, the original fill color and the current one are compared
to know if the user has changed the fill color during the edition. In
that case, the theme fill attribute and the original color are
dropped.

Some methods related to the grab bag management were added to
CellColorHandler for convenience.

Added a unit test for this attribute.

Change-Id: Ic0514ce1d2f290fb0aef5ed86327c1f03f31f20c