Improve native-code generator for (not only) Android.
Group logic from include/osl/detail/component-mapping.h has been
duplicated here for now.
The plan is to reuse this for iOS too if possible.

We don't need component-declarations.h now, which is good because
the list of implementation constructors is going to grow a lot over time.
Also, something needs to be done to avoid component-defines.h.

--constructor parameter was removed because it was not used
and also does not make sense.

__attribute__ ((visibility("default"))) is removed too.

Change-Id: I5e3f988800303d31e1d78220cbd25339bcbc482a