Related: fdo#65090 Fix crash in SwXTextTableCursor::mergeRange()
Regression from d6f896c3e61872c8380369bae7aeecf16306b0fd (fdo#48692: fix
problems with large number of table cells, 2012-09-21), the problem was
that GetSelectedBoxesCount() may return 0, and if that happens, the
prefix decrement operator set nCount to 0xffffffffffffffff, which then
resulted in a call to DeleteBox(), which did not happen before.

Thanks debug STL for pointing out this issue.

Change-Id: I9eee55f642726343c4ce9af479ccd2c4dde03b6a