Related: fdo#70624 RTF export: fix crash when header contains a page break
I don't even try to imagine how such a thing can be created and/or how
layout should handle that; but it turns out that the ODT bugdoc indeed
contains a fo:break-before="page" in the header, and we at least should
not crash on it.

Fix the problem by simple moving the init of m_pSections before the
WritePageDescTable() call.

Change-Id: If414e2f6184c58fcfd8b761d3b43fa0fc22d3903