oox: extract adjustment names by type from spec XML
The naming of adjustments is not consistent for drawingML shapes, some
of them have the first parameter as "adj1", some of them have it as
"adj". We already have a script that generates some code based on these
XML files, make it generate a function that returns the naming for each
type.

Change-Id: I2250f72854295055e1b31f2df9aec25a997db749