fdo#73215 oox: don't assume single adjustment is adj during export
So far adjustment names were either taken from the document model, or in
case there the name was empty, either "adj" was used (in case of a
single adjustment) or "adj1", "adj2", etc.

The problem is that there is no consistency here, e.g. this behavior was
correct for "cube" (single adjustment is called "adj"), but not for
"bentConnector3", where the single argument is called "adj1".

Instead of trying to guess or build a long list manually, use the new
ooxDrawingMLGetAdjNames() to write the correct names.

Change-Id: I3d609975d89c7c79f4a70c7a739cab8e01f9667f