die binfilter die die die
Euro Converter tried to obtain filters for names of binary filters that
don't exist anymore and can't be resolved hence the TypeDetection
GetByName() bailed out with an exception when calling the file picker
(e.g. when attempting to select a document), stopping in Basic
environment and effectively rendering the wizard useless.

Change-Id: Ifdef6f59abdc16bedd0f54ec317dc44af8f1e628