RTF import: handle RTF_FTNSEP and RTF_CHFTNSEP
Commit 011524d3d7f754d1e09969cabceba3f183527bb7 (fdo#68787 DOCX import:
handle when w:separator is missing for footnotes, 2013-09-04) disabled
footnote separator by default in dmapper, as the OOXML tokenizer always
provided a uFtnEdnSep in case a separator was wanted.

Let the RTF tokenizer do the same, this way we're in sync with Word
again: if RTF_CHFTNSEP is in RTF_FTNSEP, then we show the separator,
otherwise we don't.

Change-Id: I74b46c5d71227682e093695336dc9eb6fde22121