fdo#70596 - fix version dependency for linux package - 2nd try
in commit 1d1015f31de12fa486d412dea08b23f9a81a31d6 I was under the wrong
assumption that the revision is added by epm and will be always 1.

Instead the revision is passed to the epm calls in the commandline
arguments. So the upper version limit must use the "patchlevel" as the
revision. (version a.b.c.d => d is also release number)

Change-Id: I5002aef88e454561ba610b2d2774b84712da44dc