fdo#72005: Insert the border rectangle to the bottom of the stack.
This will make the border and fill shape to be drawn first, then everything
else drawn on top.

This commit may look large, but it's actually a very trivial change. The
important part is in SvxShapeGroup where new methods have been added to allow
different insertion positions for the new shapes being inserted, and have
the chart2 code make use of it to insert the fill rectangle to the bottom
rather than to the top.

Change-Id: I999160daf6fc9ce3d7e641f57b1998543df1cc4e