ooxml: Preserve shape theme attribute for solid fill
Users can select the fill color for a shape among the theme-defined
colors. This results in the following XML:

  <wps:spPr>
    ...
    <a:solidFill>
      <a:schemeClr val="accent2"/>
    </a:solidFill>
    ...
  </wps:spPr>

Now we store both the original fill color and the name of the
theme-defined color, if it exists, on the import phase. They are put
into the InteropGrabBag of the shape with the names
OriginalSolidFillClr and SpPrSolidFillSchemeClr. Additionally, we
needed to to store the decoded theme color inside StyleFillRef.

On the export phase we have to take into account several combinations
of factors:
* If the final color for the shape fill is different from the
  original color, we must ignore any theme attributes and write the
  new color.
* If the fill color is unchanged and some theme color exists, we must
  write the theme color.
* If the fill color is unchanged and no theme color exists, we must
  check if the original color matches the style-defined color. If it
  does, we must not write any <a:solidFill> tag.
* Otherwise we must write the <a:solidFill> tag with the RGB color.

The method putPropertiesToGrabBag was added to the Shape object for
convenience.

The data files for some /sd/qa/ unit tests were updated to reflect
the new properties inside the Shape InteropGrabBag.

Change-Id: If0915c5442872a8acab0a8a081f60c89c97277bd