fdo#68959: sw: fix painting of tab fill characters in small caps text
The member SwDrawTextInfo::pPos was changed from a pointer to some
externally provided Point to a real object; unfortunately there were
some functions which hold backups of the current pPos as a
reference/pointer and then restore it, and the restore is a no-op now.
So use real Points as backups.

(regression from e03819e5b4f8370d716daa1b674d50d09eee2689)

Change-Id: I8470143ec494cb194a88ddcbd12d96867974278f