Revert "coverity#704575 et al.  and clean-up."
This reverts commit 3b394be6a47482aadeee19ee659efbcfc8d4a84e, which added bogus

  if (... != 0)

checks after dynamic_casts that cannot fail, making the code now look as if it
was expected that those casts can fail and the code is prepared to handle it
(which it is not).  Silencing the false coverity reports will be addressed in a
follow-up commit.

(That commit unfortunately mixed those changes with mass cosmetic changes.)