Revert "hidpi: Blind fix - avoid double scaling by updating GetItemImage()."
The GetItemImage change causes some controls to draw wrongly and have
hit-testing problems. The change is broken because we want them to
think the bitmap is big. One day LibreOffice will use large bitmaps.

Some places call SetItemImage repeatedly, such as the font color dropdown
control and so it needs a sanity check in SetItemImage.


This reverts commit 60e9672c66eccddbb419d91cc2abef677211f84c.

Change-Id: If208543def6467ad7d19e21edf0ab11e9a32f0f4