fdo#74556: Correctly handle note captions life cycles.
When copying notes to clipboard, we don't clone captions but leave them
pointing to the original captions objects. Also, during undo and redo,
we need to clear all caption pointers to prevent them from being deleted
when the ScPostIt objects get deleted. The undo and redo of caption objects
are handled in the drawing layer afterwards.

Change-Id: I2b9cf0858dba5b3cac26db3ef501ea09779a795a