Document OutputDevice class
As part of decoupling the OutputDevice class from the Window
class, I am going through and documenting what each public function
does first. This will give a clearer indication of what Window is
responsible for, and what OutputDevice is responsible for.

Also fixed ImplDevicePixelToLogic's parameter - should not be rLogicRect,
it should be rPixelRect in the header definition.

Change-Id: I0d0b977c3d432bd08cd6d7fddc7ed878b95506d1
Reviewed-on: https://gerrit.libreoffice.org/7892
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>